requirements.txt
fastapi[standard]==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.40
python-dotenv==1.1.0
psycopg2-binary==2.9.10
asyncpg==0.30.0
python-multipart==0.0.20
passlib==1.7.4
bcrypt==4.3.0
PyJWT==2.10.1
redis==5.2.1
python-slugify==8.0.3
unidecode==1.3.7
fastapi-cache2==0.2.1
hiredis==3.0.0
alembic==1.15.1
aiofiles==23.2.1